+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 * Added prefix %prefix%
 * Fixed starting problem
 * Added language migrator
+* New permission pinata.admin.createpinata
+* New permission pinata.admin.edit
+* Added pinata create
+* Added pinata edit
+* Added pinata cratelist
 
 ### 2.6.3 update (27.03.2018)
 * Some code changes